Update project-requirement.md
Added in the Class Diagram properly, at this point a few more classes should be added, will see if anyone has capacity to help with this, then a discussion about those classes. 

Added in the State Diagram properly. This will need feedback as I was not sure how indepth to go.