Update architecture-design.md
Added in an updated class diagram, which has been slightly reworked. More changes may need to be made, but I will wait for group feedback on this before any more changes are made. (Section 4.1)

Made some minor changes to the Schedule have been made, more are to follow after getting better dates. Examples of what should be added is the powerpoint presentation that will be produced, and a breakdown of the future release. (Section 5.1)